+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### Release date ??.??.?? - Release ?.?.?
 
+- [[#173]](https://github.com/Marcin-Migdal/m-component-library/issues/173) **[PATCH]** Fix Breadcrumb covered label, implement responsiveness and tooltip on clipped crumb text,
+
 ### Release date 01.03.2025 - Release 2.1.0
 
 - [[#164]](https://github.com/Marcin-Migdal/m-component-library/issues/164) **[Minor]** Third Hot-Fix patch needed after discovering series of bugs and bad behaviors in Flaner